新建文件夹/h_pulse_mirage/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.1
aiofiles==23.2.1
python-dotenv==1.0.0
prometheus-client==0.19.0
torch==2.1.1
numpy==1.26.2
pytest==7.4.3
httpx==0.25.2
email-validator==2.1.0.post1
starlette==0.27.0
websockets==12.0
pandas==2.1.3
pillow==10.1.0 